Fix remaining compilation errors (Part 5)
- Fixed Pet domain errors:
  * Added missing error definitions: ErrFragmentMismatch, ErrFragmentTemplateNotFound, ErrInsufficientEvolutionMaterials, ErrInvalidTrainingType, ErrNoAvailableTemplates, ErrDifferentOwners, ErrInsufficientBreedingLevel
  * Fixed bondID type errors in entity.go - changed from *ActiveBond to bond.BondID
  * Fixed NewPetSkill call parameters to match function signature
  * Removed unused fmt import from value_object.go

- Fixed Minigame domain errors:
  * Added GameReward struct definition with proper fields
  * Added RewardType enum with constants
  * Fixed GameReward creation in aggregate.go with proper type conversions
  * Fixed NewMinigameAggregate call parameters
  * Commented out missing method calls (HasPlayer, CanJoin, IsFull, AddPlayer)
  * Added missing error code constants

- Fixed NPC domain errors:
  * Added missing error definitions: ErrQuestConditionsNotMet, ErrNPCHasNoShop
  * Fixed NewRelationshipChangedEvent call parameters
  * Fixed undefined variable 's' to 'ns' in service.go

- Fixed Building domain errors:
  * Removed duplicate Position and NewPosition definitions

Progress: Major domain compilation errors resolved, project much closer to compilable state.

+++ internal/interfaces/tcp/protocol/base_protocol.go
@@ -440,7 +440,7 @@ func (r *TCPRouter) RegisterHandler(msgType uint16, handler MessageHandler) {
 // HandleMessage 处理消息
 func (r *TCPRouter) HandleMessage(ctx context.Context, conn *TCPConnection, msg *Message) error {
 	r.mutex.RLock()
-	handler, exists := r.handlers[msg.Header.MessageType]
+	handler, exists := r.handlers[uint16(msg.Header.MessageType)]
 	r.mutex.RUnlock()
 
 	if !exists {
